backend/requirements.txt
# Backend Dependencies for Automated Employee Onboarding Workflow
# Python 3.11+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6  # Required for file uploads (Form data)

# CrewAI Framework
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.0
langchain-openai==0.0.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Cloud Storage
boto3==1.29.7  # AWS S3
google-cloud-storage==2.14.0  # GCP Cloud Storage (alternative)

# Document Processing
pytesseract==0.3.10  # OCR
Pillow==10.1.0  # Image processing
PyPDF2==3.0.1  # PDF processing

# Email & Notifications
sendgrid==6.11.0  # Email notifications
apns2==0.7.2  # Apple Push Notifications

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

